+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - Removes backwards compatibility for unquoted expressions (e.g. `exclude src -> dest` - use `exclude "src -> dest"` instead).
 - Adds the ability to include/exclude relationships by the tag of source of destination elements (e.g. `exclude "* -> element.tag==Tag"`).
 - Adds support for directory based includes with `!include`.
+- Adds support for element expressions to be used in the afferent/efferent coupling expression (e.g. `exclude "->element.tag==Tag"`)
 
 ## 1.17.0 (4th January 2022)